+++ pkg/core/blockchain.go
@@ -15,7 +15,6 @@ import (
 	"github.com/nspcc-dev/neo-go/pkg/config"
 	"github.com/nspcc-dev/neo-go/pkg/config/limits"
 	"github.com/nspcc-dev/neo-go/pkg/core/block"
-	"github.com/nspcc-dev/neo-go/pkg/core/blockchainer"
 	"github.com/nspcc-dev/neo-go/pkg/core/dao"
 	"github.com/nspcc-dev/neo-go/pkg/core/interop"
 	"github.com/nspcc-dev/neo-go/pkg/core/interop/contract"
@@ -179,6 +178,18 @@ type Blockchain struct {
 	events  chan bcEvent
 	subCh   chan interface{}
 	unsubCh chan interface{}
+}
+
+// StateRoot represents local state root module.
+type StateRoot interface {
+	CurrentLocalHeight() uint32
+	CurrentLocalStateRoot() util.Uint256
+	CurrentValidatedHeight() uint32
+	FindStates(root util.Uint256, prefix, start []byte, max int) ([]storage.KeyValue, error)
+	GetState(root util.Uint256, key []byte) ([]byte, error)
+	GetStateProof(root util.Uint256, key []byte) ([][]byte, error)
+	GetStateRoot(height uint32) (*state.MPTRoot, error)
+	GetLatestStateHeight(root util.Uint256) (uint32, error)
 }
 
 // bcEvent is an internal event generated by the Blockchain and then
@@ -995,8 +1006,13 @@ func (bc *Blockchain) addHeaders(verify bool, headers ...*block.Header) error {
 	return nil
 }
 
+// GetStateRoot returns state root for the given height.
+func (bc *Blockchain) GetStateRoot(height uint32) (*state.MPTRoot, error) {
+	return bc.stateRoot.GetStateRoot(height)
+}
+
 // GetStateModule returns state root service instance.
-func (bc *Blockchain) GetStateModule() blockchainer.StateRoot {
+func (bc *Blockchain) GetStateModule() StateRoot {
 	return bc.stateRoot
 }
